move serviceWorker in App
